+++ src/app/display_drivers.cpp
@@ -8,8 +8,6 @@
 
 #if DISPLAY_DRIVER == DISPLAY_DRIVER_TFT_ESPI
 #include "drivers/tft_espi_driver.cpp"
-#elif DISPLAY_DRIVER == DISPLAY_DRIVER_ST7789V2
-#include "drivers/st7789v2_driver.cpp"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX
 #include "drivers/arduino_gfx_driver.cpp"
 #elif DISPLAY_DRIVER == DISPLAY_DRIVER_ARDUINO_GFX_ST77916