+++ src/app/screens.cpp
@@ -17,7 +17,7 @@
 #include "screens/info_screen.cpp"
 #include "screens/test_screen.cpp"
 
-#if HAS_TOUCH
+#if HAS_TOUCH && LV_USE_CANVAS
 #include "screens/touch_test_screen.cpp"
 #endif